fix(rooms): drop a useless u64 conversion on the seal payload
`rooms/keys/seal`'s generated `version` is already `u64`; I copied the
`u64::from(...)` from `handle_open`, whose payload types it differently. Clippy
caught it.

Worth noting the near-miss: the obvious sed also rewrote `handle_open`'s call,
where the conversion is real. The compiler caught that one immediately, which is
the difference between a lint and a type error.

+++ vta-service/src/trust_tasks/room_group.rs
@@ -256,7 +256,7 @@ pub(super) async fn handle_seal(
         &state.room_groups_ks,
         &req.room_id,
         &req.key,
-        u64::from(req.version),
+        req.version,
         &plaintext,
     )
     .await
